refactor(all): refactoring components (moving around, fixing some basic styles, renaming) (#20)
