+++ src/app/dashboard/components/media-object.tsx
@@ -7,14 +7,20 @@ export interface MediaObjectProps {
   title: string;
 }
 
-export const MediaObject = ({ imageSrc, title, children }: PropsWithChildren<MediaObjectProps>) => (
-  <div className="mx-auto w-full rounded-md bg-gray-300 p-12 dark:bg-slate-500">
-    <div className="flex flex-col gap-8 md:flex-row">
-      <Image src={imageSrc} alt="logo" className="h-24 w-24" />
-      <section className="flex-1">
-        <h4 className="mb-4 text-xl font-medium">{title}</h4>
-        <span className="mt-1">{children}</span>
-      </section>
+export default function MediaObject({
+  imageSrc,
+  title,
+  children,
+}: PropsWithChildren<MediaObjectProps>) {
+  return (
+    <div className="mx-auto w-full rounded-md bg-gray-300 p-12 dark:bg-slate-500">
+      <div className="flex flex-col gap-8 md:flex-row">
+        <Image src={imageSrc} alt="logo" className="h-24 w-24" />
+        <section className="flex-1">
+          <h4 className="mb-4 text-xl font-medium">{title}</h4>
+          <span className="mt-1">{children}</span>
+        </section>
+      </div>
     </div>
-  </div>
-);
+  );
+}
