+++ src/app/dashboard/contacts/components/data-table/columns.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { ColumnDef } from "@tanstack/react-table";
-import { Contact } from "../../helpers/types";
+import { Contact } from "../../../helpers/types";
 import { DataTableColumnHeader } from "./data-table-column-header";
 import { DataTableRowActions } from "./data-table-row-actions";
